Update README.md
This is another unique issue that came up during weekly management meetings that needed addressing, so I created this report to analyze the missing Work Orders inputted into the EDL that would correspond to our maintenance program that was used within the company. This was all before building the Data Lake to mine different databases and create the report in PowerBI. 
By using the EDL clean data, I would input the information into this, and with functions, formatting, data-mining, and conditional formatting it would quickly show the vessel management what vessels specifically needed addressing.